feature: bind stx address through URI's query on dashboard page
