+++ src/pages/dashboard/index.js
@@ -1,0 +1,2 @@
+import Dashboard from "./[address]";
+export default Dashboard
